requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
assemblyai==0.38.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
bibtexparser==1.4.3
binaryornot==0.4.4
blinker==1.9.0
bracex==2.5.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.18.0
filetype==1.2.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
HyperPyYAML==1.2.2
idna==3.10
insanely-fast-whisper==0.0.15
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.1
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
monty==2025.3.3
more-itertools==10.6.0
mpmath==1.3.0
mrcfile==1.5.4
multidict==6.4.3
narwhals==1.37.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
omegaconf==2.3.0
openai==1.76.0
openai-whisper==20240930
optimum==1.24.0
optuna==4.3.0
orjson==3.10.16
packaging==24.2
palettable==3.3.3
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
primePy==1.3
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pymatgen==2025.4.24
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.3
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
spglib==2.6.0
SQLAlchemy==2.0.40
starlette==0.37.2
streamlit==1.44.1
streamlit-ace==0.1.1
streamlit-antd==0.8.3
streamlit-embeded==0.0.1
streamlit-molstar==0.4.21
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchao
torchmetrics==1.7.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uncertainties==3.2.3
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
wcmatch==10.0
websockets==15.0.1
yarl==1.20.0
